EPMLABSBRN-923 replace methods from UserStatisticController to UserStatisticControllerV2
